Replace constants.py with data_config.yaml (#31)
**Summary**
This PR replaces the `constants.py` file with a `data_config.yaml` file.
Dataset related settings can be defined by the user in the new yaml
file. Training specific settings were added as additional flags to the
`train_model.py` routine. All respective calls to the old files were
replaced.

**Rationale**

- Using a Yaml file for data config gives much more flexibility for
various datasets used in the community. It also facilitates the future
use of forcing and boundary datasets. In a follow-up PR the dataset
paths will be defined in the yaml file, removing the dependency on a
pre-structured `/data` folder.
- It is best practice to define user input in a yaml file, the usage of
python scripts for that purpose is not common.
- The old `constants.py` actually combined both constants and variables,
many "constants" should rather be flags to `train_models.py`
- The introduction of a new ConfigClass in `utils.py` allows for very
specific queries of the yaml and calculations based thereon. This branch
shows future possibilities of such a class
https://github.com/joeloskarsson/neural-lam/tree/feature_dataset_yaml

**Testing**
Both training and evaluation of the model were succesfully tested with
the `meps_example` dataset.

**Note**
@leifdenby Could you invite Thomas R. to this repo, in case he wanted to
give his input on the yaml file? This PR should mostly serve as a basis
for discussion. Maybe we should add more information to the yaml file as
you outline in https://github.com/mllam/mllam-data-prep. I think we
should always keep in mind how the repository will look like with
realistic boundary conditions and zarr-archives as data-input.

This PR solves parts of
https://github.com/joeloskarsson/neural-lam/issues/23

---------

Co-authored-by: Simon Adamov <simon.adamov@mailbox.org>