Change Slack invite link to new URL
Updated Slack invite link in README.

This introduces a more sustainable solution to the Slack invitie problem. Using the invite button in the App instead of the workspace admin panel, allowed me to create `never expiring` invites. 

Should the invite link still expire based on some policy change, it is now possible to simply redirect the shortened URL to a new invite link.

The choice of kutt.it over some other provider like bit.ly was due to the fact that kutt.it allows for free editing of existing urls. It further had more available short urls and is FOSS.